requirements.txt
groq
smallestai
streamlit
langchain_groq
langchain
streamlit_chat
langchain_community